fix code errors
