+++ Gemfile
@@ -3,4 +3,5 @@ source "https://rubygems.org"
 gem "byebug"
 gem "httparty"
 gem "nokogiri", "1.10.8"
+gem "rspec"
 gem "rubocop"
